perf(spark): derive owned-leaf signing pubkeys locally on coop-exit and timelock
Extends the send-path local derivation (verifying_key - operator_share) to the
other owned-leaf signing paths: cooperative exit and timelock renewal. These
signed each leaf/node with get_public_key_for_leaf, a per-item signer call that
is a network round-trip for a remote signer with a cold cache.

Promotes the derivation to a shared helper (utils::frost::derive_leaf_signing_
public_key) and points send, coop-exit, and timelock at it. Valid only for owned
leaves/nodes, where the group verifying key is the user's share plus the
operators' aggregate share; claim (incoming, mid-transfer) still asks the signer.
